requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
torch
transformers
bitsandbytes
pydantic-settings
accelerate